Report dashboard: make the Storyboard link work via on-demand HTML generation
The home dashboard rendered each session's Storyboard action as a
permanently grayed-out 'Storyboard —' label with no href. The link only
went live when a pre-generated <session>.storyboard.html happened to
exist in the logs dir, which is only produced by a manual
'trailblaze report --storyboard' run — so it was dead for every session
created through the normal server flow.

Generate the storyboard HTML on demand instead, mirroring how /report
already works:

- Extract the Playwright-free half of ReportStoryboardExporter (log →
  sections → standalone HTML page, plus the column/cell-width layout
  decisions) into a new StoryboardHtmlBuilder in trailblaze-report, so
  the server — which depends on trailblaze-report but not the host-only
  module — can reach it. The host exporter keeps the Chromium/WebP
  capture and now delegates layout to the builder; the CLI artifact is
  unchanged.
- Add a /storyboard?session=<id> endpoint that builds and serves the
  HTML on demand (404 unknown session, 400 missing param, 422 with the
  builder's actionable message for sessions with no screenshot steps).
- Point the dashboard's Storyboard link at /storyboard?session=<id> so
  it's always live; drop the dead-label branch.

Fixes #172
