+++ trailblaze-server/src/main/java/xyz/block/trailblaze/logs/server/endpoints/HomeEndpoint.kt
@@ -5,7 +5,6 @@ import io.ktor.server.response.respondText
 import io.ktor.server.routing.Routing
 import io.ktor.server.routing.get
 import io.ktor.util.toMap
-import java.io.File
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 import xyz.block.trailblaze.logs.model.SessionInfo
@@ -30,7 +29,7 @@ object HomeEndpoint {
     val rows = if (sessions.isEmpty()) {
       """<tr><td colspan="4" class="empty">No sessions yet. Run a trail to populate this list.</td></tr>"""
     } else {
-      sessions.joinToString("\n") { sessionRow(it, logsRepo) }
+      sessions.joinToString("\n") { sessionRow(it) }
     }
 
     return """
@@ -62,7 +61,6 @@ object HomeEndpoint {
           .extras { font-size: 12px; margin-top: 2px; }
           .extras a { color: #0066cc; text-decoration: none; margin-right: 12px; }
           .extras a:hover { text-decoration: underline; }
-          .extras .hint { color: #aaa; cursor: help; }
         </style>
       </head>
       <body>
@@ -91,15 +89,17 @@ object HomeEndpoint {
     """
   }
 
-  private fun sessionRow(info: SessionInfo, logsRepo: LogsRepo): String {
+  private fun sessionRow(info: SessionInfo): String {
     val sessionId = info.sessionId.value
     val (statusLabel, statusClass) = statusLabelAndClass(info.latestStatus)
     val title = htmlEscape(info.displayName)
     val when_ = htmlEscape(formatTimestamp(info))
     val duration = formatDuration(info.durationMs)
     val href = "/report?session=${urlEncode(sessionId)}"
+    // The storyboard HTML is generated on demand by StoryboardEndpoint, so this link is
+    // always live — no dependency on a pre-generated `.storyboard.html` in the logs dir.
+    val storyboardHref = "/storyboard?session=${urlEncode(sessionId)}"
     val sessionIdEscaped = htmlEscape(sessionId)
-    val storyboardExtra = storyboardLinkHtml(sessionId, logsRepo)
     return """
             <tr>
               <td><span class="status $statusClass">$statusLabel</span></td>
@@ -108,24 +108,13 @@ object HomeEndpoint {
                 <div class="meta">$sessionIdEscaped</div>
                 <div class="extras">
                   <a href="$href">Report</a>
-                  $storyboardExtra
+                  <a href="$storyboardHref">Storyboard</a>
                 </div>
               </td>
               <td class="meta nowrap">$when_</td>
               <td class="meta nowrap">$duration</td>
             </tr>
     """.trimEnd()
-  }
-
-  private fun storyboardLinkHtml(sessionId: String, logsRepo: LogsRepo): String {
-    val storyboardFile = File(logsRepo.logsDir, "reports/$sessionId.storyboard.html")
-    return if (storyboardFile.exists()) {
-      val staticPath = "/static/reports/${urlEncode(sessionId)}.storyboard.html"
-      """<a href="$staticPath">Storyboard</a>"""
-    } else {
-      val hint = htmlEscape("Run: trailblaze report --id $sessionId --storyboard")
-      """<span class="hint" title="$hint">Storyboard —</span>"""
-    }
   }
 
   private fun statusLabelAndClass(status: SessionStatus): Pair<String, String> = when (status) {
